requirements.txt
Flask~=1.1.2
WTForms>=2.3.3
Werkzeug~=1.0.1
SQLAlchemy>=1.3.16
alembic~=1.4.3
requests~=2.24.0
Flask-Assets==2.0
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
email-validator==1.1.1
python-dotenv~=0.14.0
influxdb==5.3.0
python-dateutil==2.8.1
pytz==2020.1
pandas==1.1.2
redis==3.5.3
jupyter==1.0.0

pyaml>=5.3.1
pyrosm>=0.5.3
networkx>=2.4
osmnx>=0.15.1
pyarrow>=0.17.1